feat: use json.loads
