+++ Dockerfile
@@ -20,4 +20,4 @@ RUN poetry config virtualenvs.create false \
 COPY ./scripts/create_polygons.py /app/
 COPY ./scripts/standardising.py /app/
 COPY ./scripts/aws_helper.py /app/
-COPY ./scripts/source_formatter.py /app/
+COPY ./scripts/format_source.py /app/
